% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{check_requirements}
\alias{check_requirements}
\title{check_requirements}
\usage{
check_requirements(
  run_dir,
  samp_size,
  model_list,
  ngroups,
  reference_groups,
  test_groups,
  nmfe_path,
  use_check_identifiable,
  use_simulation_data,
  simulation_data_path = NULL
)
}
\arguments{
\item{samp_size}{number of samples - must be integer}

\item{model_list}{list of model control files to be used for model averaging}

\item{ngroups}{number of groups for BE testing in data set}

\item{reference_groups}{Which groups are reference}

\item{test_groups}{Which groups are Test}

\item{nmfe_path}{nmfe??.bat}

\item{use_check_identifiable}{\itemize{
\item logical, whether to check for identifability, will check if SADDLE_RESET is in $EST
}}

\item{use_simulation_data}{logical, if the simulation will be done with a different data set than the bootstrap}

\item{simulation_data_path}{if use_simulation_data, this is the path to the data file}

\item{run.dir}{source  model files}
}
\description{
Check if all the requirements for model based BE assessment are present
}
\details{
Check to see if:
Request permission to delete the run_dir
\enumerate{
\item Does that .mod file contain ;;;;.*Start EST
\item Is the sum of reference_groups and test_groups == ngroups?
\item Any duplicates in Reference and Test groups?
\item check if data file is present
\item check if nmfe path is correct
\item check for saddle_reset if requested for check_identifiability
\item check for repeat IDs in data set
\item if use_simulation_data, see if data available
}
}
