% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BEQDF.R
\name{.get_NCA.Set.ReplicateNumber}
\alias{.get_NCA.Set.ReplicateNumber}
\title{Get NCA Set Replicate Number}
\usage{
.get_NCA.Set.ReplicateNumber(NCA.Set)
}
\arguments{
\item{NCA.Set}{The NCA set data frame.}
}
\value{
A data frame with the replicate numbers added.
}
\description{
This function calculates the replicate number for each observation in the NCA set.
}
