% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{get_base_model}
\alias{get_base_model}
\title{for each model get the control file used for the bootstrap
return a list of the models}
\usage{
get_base_model(run_dir, nmodels)
}
\arguments{
\item{nmodels}{how many models are there}
}
\description{
for each model get the control file used for the bootstrap
return a list of the models
}
\examples{
get_base_model(4)
}
