% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BEQDF.R
\name{get_BEQDF}
\alias{get_BEQDF}
\title{Get a data frame with conclusion regarding bioequivalence}
\usage{
get_BEQDF(
  NCA.Set = data.frame(),
  MetricColumn = "Cmax",
  SubjectColumn = "ID",
  TreatmentColumn = "treatment",
  SequenceColumn = "sequence",
  PeriodColumn = "period",
  RefValue = "Reference",
  alpha = 0.05,
  PartialReplicate = FALSE,
  NTID
)
}
\arguments{
\item{NCA.Set}{The NCA set data frame.}

\item{MetricColumn}{The column name for the metric of interest.
Default is \code{Cmax}}

\item{SubjectColumn}{The column name to be used as subject term.
Default is \code{ID}.}

\item{TreatmentColumn}{The column name to be used as treatment term.
Default is \code{treatment}}

\item{SequenceColumn}{The column name to be used as sequence term.
Default is \code{sequence}.}

\item{PeriodColumn}{The column name to be used as period term.
Default is \code{period}.}

\item{RefValue}{The reference value in \code{treatment} value.
Default is \code{Reference}}

\item{alpha}{The significance level for confidence intervals.
Default is 0.05.}

\item{PartialReplicate}{A logical value indicating whether partial replicate design is used.
Default is \code{FALSE}}

\item{NTID}{A logical value indicating whether RSABE criterion from Warfarin guidance should
be used.}
}
\value{
The data frame with results for the current \code{NCA.Set}.
}
\description{
This function calculates criteria required for conclusion of Bioequivalence
using the NCA set data.
}
