% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{get_block}
\alias{get_block}
\title{get_block
Get a block (e.g., $DATA) from a control file with multiple lines
return block as a single character string}
\usage{
get_block(stem, control)
}
\arguments{
\item{stem}{Block title, e.g., $DATA to look for}

\item{control}{control file text
s}
}
\description{
get_block
Get a block (e.g., $DATA) from a control file with multiple lines
return block as a single character string
}
