% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BEQDF.R
\name{get_NCA.Set.ABE}
\alias{get_NCA.Set.ABE}
\title{Get NCA Set ABE}
\usage{
get_NCA.Set.ABE(NCA.Set, alpha = 0.05)
}
\arguments{
\item{NCA.Set}{The NCA set data frame.}

\item{alpha}{The significance level for confidence intervals.
Default is 0.05.}
}
\value{
The ABE data frame with columns Ratio, lower.CL, upper.CL
as a result of mixed model analysis.
}
\description{
This function calculates the ABE (Average Bioequivalence) using the NCA set data.
}
