% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BEQDF.R
\name{get_NCA.Set.RSABE}
\alias{get_NCA.Set.RSABE}
\title{Get NCA Set RSABE}
\usage{
get_NCA.Set.RSABE(
  NCA.Set = data.frame(),
  alpha = 0.05,
  PartialReplicate = FALSE,
  NTID
)
}
\arguments{
\item{NCA.Set}{The NCA set data frame.}

\item{alpha}{The significance level for confidence intervals.
Default is 0.05.}

\item{PartialReplicate}{A logical value indicating whether partial replicate design is used.
Default is \code{FALSE}}

\item{NTID}{A logical value indicating whether RSABE criterion from Warfarin guidance should
be used.}
}
\value{
A data frame with the RSABE results.
with the columns swR, pe, critbound, VarianceCriterion
}
\description{
This function calculates the RSABE (Reference Scaled Average Bioequivalence) criterion
using the NCA set data.
}
\details{
VarianceCriterion is calculated only when \code{NTID} == \code{TRUE}.
Otherwise is is \code{NA}.
}
