% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters}
\usage{
get_parameters(
  run_dir,
  nmodels,
  samp_size,
  delta_parms,
  crash_value,
  use_check_identifiable,
  user_R_code,
  R_code_path = NULL
)
}
\arguments{
\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many samples are there}

\item{delta_parms}{criteria for identifiability test}

\item{crash_value}{value for failed calculation}

\item{use_check_identifiable}{\itemize{
\item logical, is check_identifiable to be used
}}

\item{user_R_code}{\itemize{
\item logical, whether user defined R code is to be run after the bootstrap for model averaging
}}

\item{R_code_path}{if user_R_code is TRUE, the path to the R code to be run, default is NULL}
}
\description{
read the xml file for each bootstrap/model, calculate BIC
return list of BICs (with the BEST column) and parameters for all models/samples
}
\examples{
\dontrun{
get_parameters("c:/modelaveraging", 4, 100, 0.1, 999999, TRUE)
}
}
