% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{getNCA}
\alias{getNCA}
\title{Calculate NCA Parameters from NONMEM Output}
\usage{
getNCA(
  run_dir,
  this_sample,
  NumGroups,
  reference_groups,
  test_groups,
  NCA_end_time
)
}
\arguments{
\item{run_dir}{Character string specifying the home directory where the NONMEM output files are located.}

\item{this_sample}{Integer representing the specific sample being evaluated.}

\item{NumGroups}{Integer specifying the total number of groups (e.g., 4 for an ABBA design).}

\item{reference_groups}{Numeric vector indicating the group IDs that are designated as reference.}

\item{test_groups}{Numeric vector indicating the group IDs that are designated as test.}

\item{NCA_end_time}{Numeric value indicating the end time for the AUClast and AUCinf calculations.
Calculation starts at TIME=0. A data point at TIME=0 is mandatory in the simulated study.}
}
\value{
Logical indicating the success or failure of the NCA calculation.
}
\description{
This function extracts and calculates Non-Compartmental Analysis (NCA) parameters (Cmax, AUCinf, AUClast) by group
from a NONMEM simulation output file. The output file must adhere to certain specifications,
including naming conventions and table header options.
}
\details{
The NONMEM output file should be present in the folder named \code{MBBEsimN}, where \code{N} is the simulation number.
The \verb{$TABLE} in the NONMEM file must use the \code{ONEHEADER} option and include specific columns such as \code{ID}, \code{GROUP}, \code{SEQ},
and \code{PERIOD}. For MBBE, the simulation control file, usually created by the \code{write_sim} function, will have the
table specifications like: "$TABLE ID TIME GROUP PERIOD SEQ DV EVID NOPRINT NOAPPEND FILE=OUT.DAT ONEHEADER".
This function reads the \verb{$TABLE} output from the simulation (filename \code{OUT.DAT}) and performs NCA.
}
\examples{
\dontrun{
getNCA("c:/runmodels", 1, 4, c(1, 2), c(3, 4), 72)
}

}
