% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{make_NCA_plots}
\alias{make_NCA_plots}
\title{make_NCA_plots}
\usage{
make_NCA_plots(
  BICS,
  run_dir,
  samp_size,
  nmodels,
  reference_groups,
  test_groups,
  saveplots = FALSE
)
}
\arguments{
\item{BICS}{\itemize{
\item data objects for Bayesian information Criteria
}}

\item{saveplots}{}
}
\description{
generated histograms of AUClast, AUCinf and Cmax for the BE samples
}
