% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_any_models}
\alias{run_any_models}
\title{run_any_models}
\usage{
run_any_models(nmfe_path, run_dir, nmodels, samp_size, num_parallel, BS)
}
\arguments{
\item{nmfe_path}{path to nmfe??.bat}

\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many samples are there}

\item{num_parallel}{number of models to run in parallel}

\item{BS}{logical TRUE = is bootstrap, FALSE is Monte Carlo Simulation}
}
\description{
run the bootstrap or Monte Carlo models/samples
This function can use, if available, parallel execution with multicore or multisession
The plan is set in run_mbbe
}
\examples{
\dontrun{
run.bootstrap("c:/nmfe744/util/nmfe74.bat", "c:/modelaveraging", 8)
}
}
