% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{run the bootstrap models/samples}
\usage{
run_bootstrap(
  nmfe_path,
  run_dir,
  nmodels,
  samp_size,
  numParallel = future::availableCores()
)
}
\arguments{
\item{nmfe_path}{path to nmfe??.bat}

\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many samples are there}

\item{numParallel}{how many to run in parallel, default = availableCores()}
}
\description{
run the bootstrap models/samples
}
\examples{
run.bootstrap('c:/nmfe744/util/nmfe74.bat', 'c:/modelaveraging',8)
}
