% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_mbbe_json}
\alias{run_mbbe_json}
\title{run_mbbe_json}
\usage{
run_mbbe_json(Args.json)
}
\arguments{
\item{Args.json, }{path to JSON file with arguments}
}
\value{
A list containing:
\itemize{
\item \code{Cmax_power}: Power for Cmax
\item \code{AUClast_power}: Power for AUClast
\item \code{AUCinf_power}: Power for AUCinf
\item \code{run_dir}: Directory where the function was executed
\item \code{Num_identifiable}: Number of identifiable parameters
\item \code{BICS}: Bayesian Information Criterion Scores
}
}
\description{
Runs MBBE from a json file of options e.g., calls \code{run_mbbe}
}
\examples{
\dontrun{
run_mbbe_json("Args.json")
}
}
