% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_one_model}
\alias{run_one_model}
\title{One a single NONMEM model}
\usage{
run_one_model(run_dir, nmfe_path, this_model, this_samp, BS)
}
\arguments{
\item{this_model}{which of the models used for model avering is this, integer if bootstrap, NULL if monte carlo}

\item{this_samp}{which sample (from bootstrap or Monte Carlo) is this, integer}

\item{BS}{Logical, TRUE if boostrap, FALSE if Monte Carlo}
}
\description{
One a single NONMEM model
}
\examples{
\dontrun{
 run_one_model("c:/MBBE/rundir", "c:/nm74g64/util/nmfe74.bat", 1, 1, TRUE)
}
}
