% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_simulations}
\alias{run_simulations}
\title{run all simulations}
\usage{
run_simulations(nmfe_path, run_dir, samp_size, numParallel = availableCores())
}
\arguments{
\item{nmfe_path}{path to nmfe??.bat}

\item{run_dir}{home directory}

\item{samp_size}{number of bootstrap sample for each model}

\item{numParallel}{number of models to run in parallel, default = avaiableCores()}
}
\description{
run all simulations
}
\examples{
run_simulations('c:/nmfe75/util/nmfe75.bat','c:/mbbe',8,100)
}
