% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{sample_data}
\alias{sample_data}
\title{create bootstrap samples of NONMEM data set, placed in run_dir, file names = data_sampN.csv}
\usage{
sample_data(run_dir, nmodels, samp_size)
}
\arguments{
\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many bootstrap samples, default is 100}
}
\description{
create bootstrap samples of NONMEM data set, placed in run_dir, file names = data_sampN.csv
}
\examples{
\dontrun{
sample_data("c:/modelaveraging", 100, 4)
}
}
