% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{wait_for_bs}
\alias{wait_for_bs}
\title{wait_for_bs
make array of all PIDs, check when all return Null
Stem for bootstrap is bsSamp,
Note this assume at least one run has started, could be a problem if all runs
are still running nmtran/compiler
Note that there is no time out on runs, as there models should have all run to
completion with the full data set, presumably they will be the bootstrap data set}
\usage{
wait_for_bs(nmodels, samp_size)
}
\arguments{
\item{nmodels}{number of base models}

\item{samp_size}{number of bootstrap sample for each model}
}
\description{
wait_for_bs
make array of all PIDs, check when all return Null
Stem for bootstrap is bsSamp,
Note this assume at least one run has started, could be a problem if all runs
are still running nmtran/compiler
Note that there is no time out on runs, as there models should have all run to
completion with the full data set, presumably they will be the bootstrap data set
}
\examples{
wait.for_bs(4,100)
}
