% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{wait_for_sim}
\alias{wait_for_sim}
\title{wait_for_sim
Pretty much the same as wait_for_bs, except no nmodels.
make array of all PIDs, check when all return Null
Stem for simulation is sim
Note this assume at least one run has started, could be a problem if all runs
are still running nmtran/compiler}
\usage{
wait_for_sim(samp_size)
}
\arguments{
\item{samp_size}{number of bootstrap sample for each model}
}
\description{
wait_for_sim
Pretty much the same as wait_for_bs, except no nmodels.
make array of all PIDs, check when all return Null
Stem for simulation is sim
Note this assume at least one run has started, could be a problem if all runs
are still running nmtran/compiler
}
\examples{
wait_for_sim(100)
}
