% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{calc_NCA}
\alias{calc_NCA}
\title{calc_NCA
call getNCA for each sample in 1:samp_size
and do NCA (Cmax, AUCin,AUClst, from 0 to end.time)
if check.identifability, do that with delta_parms as criteria}
\usage{
calc_NCA(
  run_dir,
  ngroups,
  reference_groups,
  test_groups,
  NCA_end_time,
  samp_size,
  numParallel = availableCores()
)
}
\arguments{
\item{run_dir, }{string home directory}

\item{reference_groups, }{arrays which groups are reference}

\item{test_groups, }{array, which groups are  test}

\item{NCA_end_time, }{numeric, end time for AUClast and AUCinf}

\item{samp_size, numeric}{integer}

\item{nGroups, }{integer how many groups, e.g., 4 for ABBA}

\item{delta_parms, }{numeric absolute difference in parameters criteria for failig identifability}
}
\description{
calc_NCA
call getNCA for each sample in 1:samp_size
and do NCA (Cmax, AUCin,AUClst, from 0 to end.time)
if check.identifability, do that with delta_parms as criteria
}
\examples{
calc_NCA('c:/runmodels',4,c(1,2),c(3,4)),72,100,0.1)
}
