% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{calc_NCA}
\alias{calc_NCA}
\title{calc_NCA}
\usage{
calc_NCA(
  run_dir,
  ngroups,
  reference_groups,
  test_groups,
  NCA_end_time,
  samp_size
)
}
\arguments{
\item{run_dir}{character;  folder path of run directory}

\item{ngroups}{numeric; integer value specifying how many groups, e.g., 4 for ABBA}

\item{reference_groups}{numeric vector; specifying which groups are reference}

\item{test_groups}{numeric vector; specifying which groups are test}

\item{NCA_end_time}{numeric; end time for AUClast and AUCinf}

\item{samp_size}{numeric; integer value specifying sample size}
}
\description{
perform NCA (Cmax, AUCin,AUClst, from 0 to NCA_end_time)
}
\details{
calc_NCA calls getNCA for each sample in 1:samp_size
Currently not parallelized
}
\examples{
\dontrun{
run_dir <- "c:/Workspace/mbbe"
ngroups <- 4
reference_groups <- c(1,2)
samp_size <- 6
test_groups <- c(3,4)
NCA_end_time <- 7
calc_NCA(run_dir, ngroups, reference_groups, test_groups, NCA_end_time, samp_size)
}
}
