% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{calc_power}
\alias{calc_power}
\title{Calculate power}
\usage{
calc_power(run_dir, samp_size, alpha, model_averaging_by, NTID)
}
\arguments{
\item{run_dir}{character, run directory}

\item{samp_size}{integer, how many samples}

\item{alpha}{numeric, range, >0 and < 1 alpha error rate}

\item{NTID}{logical, is the narrow therapeutic index drug}
}
\description{
Calculated power by doing EMA standards statistics on each Monte Carlo simulation, then counting the number that pass BE
}
\examples{
\dontrun{
nca_results <- system.file(package = "mbbe", "nca_results")
calc_power(nca_results, 200, 0.05, FALSE)
}
}
