% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{check_identifiable}
\alias{check_identifiable}
\title{read .lst file, find out where saddle reset occurs then get parameter before reset from .ext file (Pre.parms)
compare with final parameters (Post.parms), see if any differ by delta_parms
open .lst}
\usage{
check_identifiable(run_dir, this_model, this_sample, delta_parms, nparms)
}
\arguments{
\item{run_dir}{home directory}

\item{this_model}{integer}

\item{this_sample}{integer}

\item{delta_parms}{absolute difference in parameters criteria for failing identifability}

\item{nparms}{number of parameters in the .ext file (can be less thn the total)}
}
\description{
read .lst file, find out where saddle reset occurs then get parameter before reset from .ext file (Pre.parms)
compare with final parameters (Post.parms), see if any differ by delta_parms
open .lst
}
\examples{
\dontrun{
check_identifiable("c:/runmodels", 1, 1, 0.1)
}

}
