% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{check_requirements}
\alias{check_requirements}
\title{Check Requirements for Model-Based BE Assessment}
\usage{
check_requirements(
  run_dir,
  samp_size,
  model_list,
  ngroups,
  reference_groups,
  test_groups,
  nmfe_path,
  use_check_identifiable,
  simulation_data_path,
  user_R_code,
  R_code_path = NULL
)
}
\arguments{
\item{run_dir}{Character string specifying the directory containing the source model files.}

\item{samp_size}{Integer specifying the number of samples.}

\item{model_list}{List of character strings representing the model control files to be utilized for model averaging.}

\item{ngroups}{Integer representing the number of groups for BE testing in the dataset.}

\item{reference_groups}{Character vector indicating the groups classified as "reference".}

\item{test_groups}{Character vector indicating the groups classified as "test".}

\item{nmfe_path}{Character string indicating the path to the nmfe batch file (e.g., nmfe?.bat).}

\item{use_check_identifiable}{Logical. Determines whether to verify the identifiability by checking if \code{SADDLE_RESET} is present in $EST.}

\item{simulation_data_path}{Character string specifying the path to the data file used for simulation.}

\item{user_R_code}{Logical. Should user-defined R code be executed post the NONMEM run for bootstrap?}

\item{R_code_path}{Optional. Character string indicating the path to the user-defined R code to be executed post the NONMEM bootstrap. This returns a penalty value to be appended to the BIC.}
}
\value{
\code{TRUE} if all requirements pass and \code{FALSE} if any failures.
}
\description{
This function verifies the necessary requirements for conducting a model-based bioequivalence (BE) assessment.
}
\details{
The function checks for the following:
\enumerate{
\item Presence of the \code{.mod} file containing \verb{;;;;.*Start EST}.
\item Verification that the sum of reference and test groups equals \code{ngroups}.
\item Absence of duplicate values in both reference and test groups.
\item Presence of the specified data file.
\item Validity of the provided \code{nmfe_path}.
\item (If requested) Verification for \code{SADDLE_RESET} in the context of identifiability.
\item Absence of repeat IDs in the dataset.
\item Availability of simulation data.
}
}
