% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{check_requirements}
\alias{check_requirements}
\title{check_requirements
2. Does that .mod file contain ;;;;.*Start EST
3. is the sum of reference_groups and test_groups == ngroups?
4. Any duplicates in Reference and Test groups?
5. check if data file is present
6. check if nmfe path is correct
7. check for saddle_reset if requested
8. check for repeat IDs in data set
9. if use_simulation_data, see if data available}
\usage{
check_requirements(
  run_dir,
  samp_size,
  model_list,
  ngroups,
  reference_groups,
  test_groups,
  nmfe_path,
  use_check_identifiable,
  use_simulation_data,
  simulation_data_path = NULL
)
}
\arguments{
\item{ngroups}{number of groups in data set}

\item{reference_groups}{Which groups are reference}

\item{test_groups}{Which groups are Test}

\item{nmfe_path}{nmfe??.bat}

\item{use_check_identifiable}{\itemize{
\item logical, whether to check for identifability, will check if SADDLE_RESET is in $EST
}}

\item{use_simulation_data}{logical, if the simulation will be done with a different data set than the bootstrap}

\item{simulation_data_path, }{if use_simulation_data, this is the path to the data file}

\item{source.dir}{source  model files}

\item{nmodels}{number of models}
}
\description{
check_requirements
2. Does that .mod file contain ;;;;.*Start EST
3. is the sum of reference_groups and test_groups == ngroups?
4. Any duplicates in Reference and Test groups?
5. check if data file is present
6. check if nmfe path is correct
7. check for saddle_reset if requested
8. check for repeat IDs in data set
9. if use_simulation_data, see if data available
}
