% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{copy_model_files}
\alias{copy_model_files}
\title{copy_model_files}
\usage{
copy_model_files(model_source, run_dir)
}
\arguments{
\item{run_dir}{Folder where models are to be run}

\item{list}{of NONMEM model files}
}
\value{
nmodels how many models are there
}
\description{
copy NONMEM control files from a source to a run directory
no change in control files
}
\examples{
\dontrun{
copy_model_files("c:/mbbe/models/model1.mod","c:/mbbe/models/model2.mod"), "c:/mbbe/run")
}
}
