% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{copy_model_files}
\alias{copy_model_files}
\title{copy_model_files
copy NONMEM model files from a source to a run directory
Need to look at this for OS portability??}
\usage{
copy_model_files(model_source, run_dir)
}
\arguments{
\item{model_source}{folder name were subfolders (model1..modelnmodels) with model files can be found. There must be exactly one .mod file (with any stem) in the folder}

\item{run_dir}{Folder where models are to be run}
}
\value{
nmodels how many models are there
}
\description{
copy_model_files
copy NONMEM model files from a source to a run directory
Need to look at this for OS portability??
}
\examples{
\dontrun{
copy_model_files("c:/modelaveraging", "c:/modelaveraging/run", 4)
}
}
