% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{get_parameters}
\alias{get_parameters}
\title{get_parameters
read the xml file for each bootstrap/model, calculate BIC
return list of BICs (with the BEST column) and parameters for all models/samples}
\usage{
get_parameters(
  run_dir,
  nmodels,
  samp_size,
  delta_parms,
  crash_value,
  use_check_identifiable
)
}
\arguments{
\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many samples are there}

\item{delta_parms}{criteria for identifability test}

\item{crash_value}{value for failed calculation}
}
\description{
get_parameters
read the xml file for each bootstrap/model, calculate BIC
return list of BICs (with the BEST column) and parameters for all models/samples
}
\examples{
get_parameters('c:/modelaveraging',4,100,0.1)
}
