% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{getNCA}
\alias{getNCA}
\title{read $TABLE output from simulation (file name out.dat)
and do NCA (Cmax, AUCin,AUClst, from 0 to end.time)
if check.identifability, do that with delta_parms as criteria}
\usage{
getNCA(
  run_dir,
  this_sample,
  NumGroups,
  reference_groups,
  test_groups,
  NCA_end_time
)
}
\arguments{
\item{run_dir}{home directory}

\item{this_sample}{integer}

\item{NumGroups}{integer how many groups, e.g., 4 for ABBA}

\item{reference_groups}{list of arrays, which groups are reference}

\item{test_groups}{list of arrays, which groups are  test}

\item{NCA_end_time}{end time for AUClast and AUCinf}
}
\description{
read $TABLE output from simulation (file name out.dat)
and do NCA (Cmax, AUCin,AUClst, from 0 to end.time)
if check.identifability, do that with delta_parms as criteria
}
\examples{
getNCA('c:/runmodels',1,4,c(1,2),c(3,4)),72,0.1)
}
