% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{getNCA}
\alias{getNCA}
\title{getNCA}
\usage{
getNCA(
  run_dir,
  this_sample,
  NumGroups,
  reference_groups,
  test_groups,
  NCA_end_time
)
}
\arguments{
\item{run_dir}{home directory}

\item{this_sample}{integer}

\item{NumGroups}{integer how many groups, e.g., 4 for ABBA}

\item{reference_groups}{list of arrays, which groups are reference}

\item{test_groups}{list of arrays, which groups are test}

\item{NCA_end_time}{numeric, end time for AUClast and AUCinf, calculation starts at TIME=0 (TIME=0 data point is required in simulated study)}
}
\value{
Logical
}
\description{
Calculated NCA parameters (Cmax, AUCinf, AUClast) by group from NONMEM output file
NONMEM output file MUST be in the folder MBBEsimN where N is the simulation number
$TABLE must include the ONEHEADER option
The simulation must include an observation at T=
The $TABLE file output by the simulation must include:
ID
GROUP
SEQ
PERIOD
Note that for MBBE, the simulation control file is written by the write_sim function and will include:
"$TABLE ID TIME GROUP PERIOD SEQ DV EVID NOPRINT NOAPPEND FILE=OUT.DAT ONEHEADER")
reads $TABLE output from simulation (file name out.dat)
and do NCA (Cmax, AUCin,AUClst, from 0 to end.time)
if check.identifiability, do that with delta_parms as criteria
}
\examples{
\dontrun{
getNCA("c:/runmodels", 1, 4, c(1, 2), c(3, 4), 72, 0.1)
}
}
