% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{make_NCA_plots}
\alias{make_NCA_plots}
\title{Title}
\usage{
make_NCA_plots(
  BICS,
  run_dir,
  samp_size,
  nmodels,
  reference_groups,
  test_groups,
  saveplots = FALSE
)
}
\arguments{
\item{saveplots}{}
}
\description{
Title
}
