% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_any_models}
\alias{run_any_models}
\title{run the bootstrap or monte carlo models/samples}
\usage{
run_any_models(nmfe_path, run_dir, nmodels, samp_size, BS, plan, num_parallel)
}
\arguments{
\item{nmfe_path}{path to nmfe??.bat}

\item{run_dir}{Folder where models are to be run}

\item{nmodels}{how many models are there}

\item{samp_size}{how many samples are there}

\item{BS}{logical TRUE = is bootstrap, FALSE is Monte Carlo Simulation}

\item{plan}{\itemize{
\item future plan, one of sequential, multicore, multisession
}}

\item{num_parallel}{how many to run in parallel, default = availableCores()}
}
\description{
run the bootstrap or monte carlo models/samples
}
\examples{
\dontrun{
run.bootstrap("c:/nmfe744/util/nmfe74.bat", "c:/modelaveraging", 8)
}
}
