% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_mbbe}
\alias{run_mbbe}
\title{run_mbbe}
\usage{
run_mbbe(
  crash_value,
  ngroups,
  reference_groups,
  test_groups,
  num_parallel,
  samp_size,
  run_dir,
  model_source,
  nmfe_path,
  delta_parms,
  use_check_identifiable,
  NCA_end_time,
  rndseed,
  use_simulation_data,
  simulation_data_path,
  plan = c("multisession", "sequential", "multicore"),
  alpha_error = 0.05,
  NTID,
  model_averaging_by,
  save_output = FALSE
)
}
\arguments{
\item{simulation_data_path}{}
}
\description{
run_mbbe
}
