% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_mbbe}
\alias{run_mbbe}
\title{run_mbbe}
\usage{
run_mbbe(
  crash_value,
  ngroups,
  reference_groups,
  test_groups,
  num_parallel,
  samp_size,
  run_dir,
  model_source,
  nmfe_path,
  delta_parms,
  use_check_identifiable,
  NCA_end_time,
  rndseed,
  simulation_data_path,
  plan = c("multisession", "sequential", "multicore"),
  alpha_error = 0.05,
  NTID,
  model_averaging_by,
  user_R_code,
  R_code_path
)
}
\arguments{
\item{crash_value}{\itemize{
\item value to be returned for BIC in models that crash, either in bootstrap or simulation
}}

\item{ngroups}{\itemize{
\item number of groups in simulated data, e.g., ABBA design has 4 groups
}}

\item{reference_groups}{\itemize{
\item which of the groups are reference formulation e.g., ABBA design might be c(2,3)
}}

\item{test_groups}{\itemize{
\item which of the groups are test formulation e.g., ABBA design might be c(1,4)
}}

\item{num_parallel}{\itemize{
\item number of NONMEM (bootstrap and simulation) to be run in parallel
}}

\item{samp_size}{\itemize{
\item size of bootstrap and simulation sample
}}

\item{run_dir}{\itemize{
\item where to run NONMEM
}}

\item{model_source}{\itemize{
\item NONMEM control files to be used/averaged over
}}

\item{nmfe_path}{\itemize{
\item path to nmfe??.bat
}}

\item{delta_parms}{\itemize{
\item difference in parameters used to define identifiability
}}

\item{use_check_identifiable}{\itemize{
\item logical, whether to use identifiability (defined by Aoki (https://www.page-meeting.org/default.asp?abstract=5951))
}}

\item{NCA_end_time}{\itemize{
\item NCA calculation will start at 0, end at NCA_end_time
}}

\item{rndseed}{\itemize{
\item random seed
}}

\item{simulation_data_path}{\itemize{
\item path to simulation data set
}}

\item{plan}{\itemize{
\item the parallel exection plan, one of "multisession", "sequential", "multicore"
}}

\item{alpha_error}{alpha error for statistical testing}

\item{NTID}{logical , whether the drug is a narrow therapeutic index drug}

\item{model_averaging_by}{one "study", "subject"}

\item{user_R_code}{\itemize{
\item logical, whether user provide code us used for the penalty for selectin models
}}

\item{R_code_path}{\itemize{
\item if user_R_code, required path to R code
}}
}
\value{
Cmax_power,
AUClast_power,
AUCinf_power,
run_dir,
Num_identifiable,
BICS
}
\description{
Runs MBBE, typically called by run_mbbe_json, using a json file that includes the required options
}
\details{
{
Typically called from run_mbbe_json}
}
