% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_mbbe_json}
\alias{run_mbbe_json}
\title{run_mbbe_json}
\usage{
run_mbbe_json(Args.json)
}
\arguments{
\item{Args.json, }{path to JSON file with arguments}
}
\description{
Runs MBBE from a json file of optoins
Calls run_mbbe
}
\examples{
\dontrun{
run_mbbe_json(Args.json)
}
}
