% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{run_mbbe_json}
\alias{run_mbbe_json}
\title{run_mbbe_json, reads json file, calls run_mbbe}
\usage{
run_mbbe_json(Args.json)
}
\arguments{
\item{Args.json, }{path to JSON file with arguments}
}
\description{
run_mbbe_json, reads json file, calls run_mbbe
}
\examples{
run_mbbe_json(Args.json)
}
