% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBBE.R
\name{write_sim_controls}
\alias{write_sim_controls}
\title{Edits the best based model for each sample, replaces the original parameters with the bootstrap parameters
and adds the $SIM and $TABLE}
\usage{
write_sim_controls(
  run_dir,
  parms,
  base_models,
  samp_size,
  use_simulation_data,
  simulation_data_path = NULL
)
}
\arguments{
\item{run_dir}{Folder where models are to be run}

\item{parms}{list that include BICs and parameters for each sample/model}

\item{base_models}{list of the text in each model used for the bootstrap}

\item{samp_size}{how many samples are there}

\item{use_simulation_data}{logical - TRUE is a file other than the bootstrap data file is to be used for simulation}

\item{simulation_data_path}{if use_simulation_data==TRUE, the path to the file tto be used for simulation}
}
\description{
Edits the best based model for each sample, replaces the original parameters with the bootstrap parameters
and adds the $SIM and $TABLE
}
\examples{
\dontrun{
write_sim_controls("c:/modelaveraging", parms, base_models, 100)
}
}
